Move japanese era data struct into icu_calendar
