+++ provider/cldr/Cargo.toml
@@ -42,6 +42,7 @@ icu_plurals = { version = "0.4", path = "../../components/plurals" }
 icu_datetime = { version = "0.4", path = "../../components/datetime", features = ["provider_transform_internals"] }
 icu_locale_canonicalizer = { version = "0.4", path = "../../components/locale_canonicalizer" }
 icu_decimal = { version = "0.4", path = "../../components/decimal" }
+icu_calendar = { version = "0.4", path = "../../components/calendar", features = ["provider_serde"] }
 icu_list = { version = "*", path = "../../experimental/list_formatter", features = ["provider_transform_internals"]}
 itertools = "0.10"
 json = "0.12"
